[psu] share the PSU line pattern across tests
